SimpleRouter version 1.0 - a library designed to manage routing in a .NET application. This library includes the definition of routes, a router for managing navigation between routes, and events for handling route changes.
RouteChangingEventArgs and RouteChangedEventArgs - used to provide data for route changing events.

IRoute interface - defines the properties that a route should have, and the RouteFactory delegate defines a method that creates a route.

IRouter interface - defines the properties and methods that a router should have. This includes methods for navigating to a route, navigating back, and events for when the route is changing or has changed.

IRouterHost interface - defines an entity that will be responsible for managing the router.

Router - implements the IRouter interface. This class contains the implementation of the router's properties and methods.